requirements.txt
asgiref==3.6.0
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==40.0.2
daphne==4.0.0
Django==4.2.1
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
packaging==23.1
psycopg2==2.9.7
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.1.1
python-decouple==3.8
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
zope.interface==6.0
whitenoise==6.4.0
django-cors-headers==4.2.0